requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# HTTP client for API calls
httpx==0.25.2
aiohttp==3.9.1
openai>=1.3.0
python-dotenv>=1.0.0

# File watching and async support
watchdog==3.0.0
asyncio-mqtt>=0.11.0
python-multipart>=0.0.6

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Optional: For advanced features
redis==5.0.1  # If using Redis for caching
celery==5.3.4  # If using Celery for background tasks
